코드 리팩토링 및 프로토콜 메시지 업데이트
- `Action.h`, `Functions.h`, `PhysicsUtils.h` 파일에서 `struct`와 `class` 변경
- `Engine.vcxproj` 파일에 `/IGNORE:4006` 옵션 추가
- 여러 파일에서 타입 캐스팅 추가 및 경고 비활성화
- `ProtoInclude.h` 파일에 `#include "Connect.pb.h"` 추가
- `Dll Library.props` 파일의 `OutDir` 경로 수정
- 불필요한 헤더 파일 삭제
- `Description.h` 파일의 초기값 수정
- `Connect.proto`, `Move.proto`, `Play.proto` 파일의 메시지 수정 및 추가

+++ 4Q_Solution/Server/PacketDispatcher.cpp
@@ -23,7 +23,7 @@ void PacketDispatcher::SaveRecvPacket(StreamBuffer* recvData, SessionID sid)
 
 		Packet packet;
 		recvData->Read(PtrCast(char*, &packet), (int)header._packetSize);
-		packet.sessionId = sid;
+		packet.sessionId = static_cast<short>(sid);
 		//printf("[Dispatcher::SaveRecvPacket] Packet Get. Size : %d\n", (int)header._packetSize);
 		Lock lock(_recvMtx);
 		_saveRecvContainer.push(std::move(packet));
@@ -104,10 +104,10 @@ void PacketDispatcher::SaveBroadCastPacket(std::string data, short packetId, lon
 
 void PacketDispatcher::MakePacket(Packet& packet, std::string data, SessionID sid, short packetId, long dataSize, int serialNum)
 {
-	packet.sessionId = sid;
+	packet.sessionId = static_cast<short>(sid);
 	packet._packetId = packetId;
-	packet._serialNumber = serialNum;
+	packet._serialNumber = static_cast<short>(serialNum);
 
 	memcpy(packet._data, data.c_str(), dataSize);
-	packet._packetSize = sizeof(PacketHeader) + dataSize;
+	packet._packetSize = static_cast<short>(sizeof(PacketHeader) + dataSize);
 }
